refactor: update component imports to improve structure and organization
