+++ src/app/[lng]/admin/page.tsx
@@ -1,10 +1,11 @@
-import React, { useState } from "react";
+import React from "react";
 
-import AdminPage from "@/components/AdminPage";
-import Researches from "@/components/Researches";
+import AdminPage from "@/components/admin/admin-page";
+import Researches from "@/components/research/research";
 import Departments from "@/components/departments/Departments";
 import { PageProps } from "@/types";
-import GalleryPage from "@/components/pages/gallery-page-component";
+import GalleryPage from "@/components/gallery/gallery-page-component";
+import Standard from "@/components/standards/standards";
 
 const Admin = ({ params }: PageProps) => {
   const { lng } = params;
@@ -14,10 +15,17 @@ const Admin = ({ params }: PageProps) => {
       <AdminPage
         pages={{
           researches: <Researches key="research" />,
+          standards: <Standard key="standard" />,
           departments: <Departments key="department" />,
           gallery: <GalleryPage key="gallery" />,
         }}
-        tabs={["Researches", "Departments", "Gallery", "Home Page"]}
+        tabs={[
+          "Researches",
+          "Standards",
+          "Departments",
+          "Gallery",
+          "Home Page",
+        ]}
         lng={lng}
       />
     </>
