+++ src/app/[lng]/news/[slug]/page.tsx
@@ -2,12 +2,12 @@
 import React, { useEffect } from "react";
 import { PostOrPage } from "@tryghost/content-api";
 import ImageWithTextOverlay from "@/components/image-overlay";
-import SocialLinks from "@/components/social-icons";
+import SocialLinks from "@/components/ui/social-icons";
 import useLoading from "@/hooks/useLoading";
 import Image from "next/image";
 import { getSinglePost } from "@/app/ghost/posts";
 import dayjs from "dayjs";
-import SafeHTML from "@/components/safe-html";
+import SafeHTML from "@/components/ui/safe-html";
 import Link from "next/link";
 
 export const NewsPage = ({ params }: { params: { slug: string } }) => {
